pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aibot"
version = "0.1.0"
description = "NoneBot2 + NapCatQQ (OneBot v11)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nonebot2[fastapi,websockets]>=2.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "pillow>=10.0.0",
]

[tool.setuptools]
py-modules = ["bot", "openclaw_memory"]

[tool.nonebot]
plugins = []
plugin_dirs = ["plugins"]
